Remove unfair TrustedString wrapping in prod dataset
Discovered that pslog.TrustedString crept into the prod dataset
prep skewing the results quite heavily in pslog's favor. This
was an AI regression from previously when the AI agent did the
same thing. It was removed, then it came back. pslog is
fast, but not that fast w/o further optimizations. This is
a classic case of an AI artifact, improving-at-all-costs.
Need to monitor this more in the future as the v0.1.0 and v0.2.0
essentially lie about performance... however, now it's possible
to improve further for real which was not really possible before
as no iteration led to an improved as we had already reached
the horizon of what could be reasonably possible.